# Use stack buffer instead of dynamic allocation when save/load rdb