I find we still use some dynamic memory allocations in save or load RDB, for lzf compressed buffer, we always allocate when needed. I think we can use stack buffer to reduce dynamic allocation if the needed space is not big.

As we know, the maximum consumption in saving/loading RDB is `lzf compress/uncompress`, in fact, i don't find this commit makes RDB loading/saving use less time, but in some scene, after this commit, redis bgsave child process uses less COW memory.

In my test, redis load a rdb that may have some different size key/value, and then use benchmark `redis-benchmark -t set -d 4096 -r 1000000 -n 1000000000` to continue writing redis.
Before
```
29336:M 20 Apr 2021 17:57:43.067 * 10000 changes in 60 seconds. Saving...
29336:M 20 Apr 2021 17:57:43.227 * Background saving started by pid 100425
100425:C 20 Apr 2021 18:00:15.781 * DB saved on disk
100425:C 20 Apr 2021 18:00:15.957 * RDB: 5874 MB of memory used by copy-on-write
29336:M 20 Apr 2021 18:00:16.599 * Background saving terminated with success
```
Now
```
49808:M 20 Apr 2021 17:34:30.009 * 10000 changes in 60 seconds. Saving...
49808:M 20 Apr 2021 17:34:30.164 * Background saving started by pid 47585
47585:C 20 Apr 2021 17:37:00.153 * DB saved on disk
47585:C 20 Apr 2021 17:37:00.333 * RDB: 4947 MB of memory used by copy-on-write
```

For loading RDB, this commit doesn't decrease time, i am not sure we should do that, but it should be good to reduce memory dynamic allocation.