@oranagra 2021-04-20
@ShooterIT i'm not certain this is really CoW... I remind you that what we're actually measuring is private dirty bytes (not CoW).
so if the allocator allocated something and didn't return it yet to the OS (keeping it for future use), it'll seem like CoW.
This is nearly 1gb, which seems a lot, but maybe due to each allocation falling into a different bin, it eventually consumes a lot.
one way to check might be to flush the jemalloc tcache and purge the arenas before measuring the "CoW".

if it is really CoW, which is caused by the fact the OS has to copy-on-write these "new allocations", since they're using the same pages the main process used before forking, maybe there's a better solution for that (one that's not bound only to small allocations like you did).
maybe we should purge the allocator before calling fork, so that the unused pages it keeps for future use, won't be shared between the two processes.