@ShooterIT 2021-04-20
Thanks @oranagra i tend to believe this memory is not COW, but such small change makes` private dirty bytes` less than before, so i make a PR. Yes, the memory allocator doesn't timely return memory to OS because of performance, but we always face this scene.

> maybe we should purge the allocator before calling fork, so that the unused pages it keeps for future use, won't be shared between the two processes.

do you mean we call `jemalloc_purge()` before calling `fork`? from comments, but it will block server.