@oranagra 2021-04-20
> but such small change makes private dirty bytes less than before

if we test it (by purging before measuring COW) we may find out that it doesn't really do anything..  i.e. in your test it does because it reduces the number of temporary allocations the child does. but we may realize that it's just a measurement error rather than a real improvement.

yes, i mean calling `jemalloc_purge`, and possibly first do mallctl with `thread.tcache.flush` (see memeffieciency.tcl) too.
i would suggest you test both of these theories (seeing what would be the difference if these are called before fork, and also if called before measuring cow in the child).