@ShooterIT 2021-04-21
copy that, but i still want to say we always face that allocator keep allocations and doesn't return to OS timely. I will do some tests as you said, i think it may meet your expectations.